Finished specs for metrics
