+++ Guardfile
@@ -8,9 +8,9 @@ guard :minitest do
   watch(%r{^test/test_helper\.rb})      { 'test' }
 
   # with Minitest::Spec
-  # watch(%r{^spec/(.*)_spec\.rb})
-  # watch(%r{^lib/(.+)\.rb})         { |m| "spec/#{m[1]}_spec.rb" }
-  # watch(%r{^spec/spec_helper\.rb}) { 'spec' }
+  watch(%r{^spec/(.*)_spec\.rb})
+  watch(%r{^lib/(.+)\.rb})         { |m| "spec/#{m[1]}_spec.rb" }
+  watch(%r{^spec/spec_helper\.rb}) { 'spec' }
 
   # Rails 4
   # watch(%r{^test/test_helper\.rb}) { 'test' }
