+++ Rakefile
@@ -1,5 +1,5 @@
 task :spec do
-  Dir.glob('./spec/*_spec.rb') do |f|
+  Dir.glob('./spec/mm/*_spec.rb') do |f|
     ruby f
   end
 end
